secured endpoints

+++ pizza-order-rest/src/main/java/hu/elte/pizzaorder/controllers/OrderController.java
@@ -1,8 +1,9 @@
 package hu.elte.pizzaorder.controllers;
 
 import hu.elte.pizzaorder.entities.Order;
-import hu.elte.pizzaorder.entities.Order;
 import hu.elte.pizzaorder.repositories.OrderRepository;
+import hu.elte.pizzaorder.entities.User;
+import hu.elte.pizzaorder.security.AuthenticatedUser;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -20,6 +21,9 @@ import org.springframework.web.bind.annotation.RestController;
 public class OrderController {
     @Autowired
     private OrderRepository orderRepository;
+    
+    @Autowired
+    private AuthenticatedUser authenticatedUser;
     
     @GetMapping("")
     public ResponseEntity<Iterable<Order>> getAll() {
@@ -44,8 +48,10 @@ public class OrderController {
     
     @PutMapping("/{id}")
     public ResponseEntity<Order> update(@PathVariable Integer id, @RequestBody Order order) {
+        User user = authenticatedUser.getUser();
+        User.Role role = user.getRole();
         Optional<Order> oOrder = orderRepository.findById(id);
-        if (oOrder.isPresent()) {
+        if (oOrder.isPresent() && role.equals(User.Role.ADMIN)) {
             order.setId(id);
             return ResponseEntity.ok(orderRepository.save(order));
         } else {
@@ -55,8 +61,10 @@ public class OrderController {
     
     @DeleteMapping("/{id}")
     public ResponseEntity<Order> delete(@PathVariable Integer id) {
+        User user = authenticatedUser.getUser();
+        User.Role role = user.getRole();
         Optional<Order> oOrder = orderRepository.findById(id);
-        if (oOrder.isPresent()) {
+        if (oOrder.isPresent() && role.equals(User.Role.ADMIN)) {
             orderRepository.deleteById(id);
             return ResponseEntity.ok().build();
         } else {
